add camera

+++ src/Sprite.hpp
@@ -18,12 +18,12 @@ class Sprite{
         void setPosition(float x, float y);
         void setSize(float w, float h);
 
-        inline math::Vector2<int> getPosition(){
-            return math::Vector2<int>(Size_Pos.x, Size_Pos.y);
+        inline math::Vector2<float> getPosition(){
+            return math::Vector2<float>(Size_Pos.x, Size_Pos.y);
         };
 
-        inline math::Vector2<int> getSize(){
-            return math::Vector2<int>(Size_Pos.w, Size_Pos.h);
+        inline math::Vector2<float> getSize(){
+            return math::Vector2<float>(Size_Pos.w, Size_Pos.h);
         };
 
         inline void setTexture(SDL_Texture* T){
